+++ src/meetingsession/DefaultMeetingSession.ts
@@ -12,6 +12,11 @@ import ContentShareMediaStreamBroker from '../contentsharecontroller/ContentShar
 import DefaultContentShareController from '../contentsharecontroller/DefaultContentShareController';
 import Destroyable, { isDestroyable } from '../destroyable/Destroyable';
 import DeviceController from '../devicecontroller/DeviceController';
+import EventIngestionConfiguration from '../eventingestionconfiguration/EventIngestionConfiguration';
+import DefaultMeetingEventReporter from '../eventreporter/DefaultMeetingEventReporter';
+import EventReporter from '../eventreporter/EventReporter';
+import EventsClientConfiguration from '../eventsclientconfiguration/EventsClientConfiguration';
+import MeetingEventsClientConfiguration from '../eventsclientconfiguration/MeetingEventsClientConfiguration';
 import Logger from '../logger/Logger';
 import DeviceControllerBasedMediaStreamBroker from '../mediastreambroker/DeviceControllerBasedMediaStreamBroker';
 import DefaultReconnectController from '../reconnectcontroller/DefaultReconnectController';
@@ -26,17 +31,20 @@ export default class DefaultMeetingSession implements MeetingSession, Destroyabl
   private contentShareController: ContentShareController;
   private _deviceController: DeviceController;
   private audioVideoFacade: AudioVideoFacade;
+  private _eventReporter: EventReporter;
 
   constructor(
     configuration: MeetingSessionConfiguration,
     logger: Logger,
-    deviceController: DeviceControllerBasedMediaStreamBroker
+    deviceController: DeviceControllerBasedMediaStreamBroker,
+    eventReporter?: EventReporter
   ) {
     this._configuration = configuration;
     this._logger = logger;
 
     this.checkBrowserSupportAndFeatureConfiguration();
 
+    this.setupEventReporter(configuration, logger, eventReporter);
     this._deviceController = deviceController;
     this.audioVideoController = new DefaultAudioVideoController(
       this._configuration,
@@ -50,7 +58,8 @@ export default class DefaultMeetingSession implements MeetingSession, Destroyabl
           this._configuration.reconnectShortBackOffMs,
           this._configuration.reconnectLongBackOffMs
         )
-      )
+      ),
+      this._eventReporter
     );
     deviceController.bindToAudioVideoController(this.audioVideoController);
     const contentShareMediaStreamBroker = new ContentShareMediaStreamBroker(this._logger);
@@ -104,6 +113,10 @@ export default class DefaultMeetingSession implements MeetingSession, Destroyabl
     return this._deviceController;
   }
 
+  get eventReporter(): EventReporter {
+    return this._eventReporter;
+  }
+
   /**
    * Clean up this instance and resources that it created.
    *
@@ -116,6 +129,9 @@ export default class DefaultMeetingSession implements MeetingSession, Destroyabl
     if (isDestroyable(this.audioVideoController)) {
       await this.audioVideoController.destroy();
     }
+    if (isDestroyable(this.eventReporter)) {
+      await this.eventReporter.destroy();
+    }
 
     this._logger = undefined;
     this._configuration = undefined;
@@ -123,6 +139,36 @@ export default class DefaultMeetingSession implements MeetingSession, Destroyabl
     this.audioVideoFacade = undefined;
     this.audioVideoController = undefined;
     this.contentShareController = undefined;
+    this._eventReporter = undefined;
+  }
+
+  private setupEventReporter(
+    configuration: MeetingSessionConfiguration,
+    logger: Logger,
+    eventReporter?: EventReporter
+  ): void {
+    if (eventReporter) {
+      this._eventReporter = eventReporter;
+    } else {
+      const eventIngestionURL = configuration.urls.eventIngestionURL;
+      if (eventIngestionURL) {
+        this.logger.info(`Event ingestion URL is present in the configuration`);
+        const {
+          meetingId,
+          credentials: { attendeeId, joinToken },
+        } = configuration;
+        const meetingEventsClientConfiguration: EventsClientConfiguration = new MeetingEventsClientConfiguration(
+          meetingId,
+          attendeeId,
+          joinToken
+        );
+        const eventIngestionConfiguration = new EventIngestionConfiguration(
+          meetingEventsClientConfiguration,
+          eventIngestionURL
+        );
+        this._eventReporter = new DefaultMeetingEventReporter(eventIngestionConfiguration, logger);
+      }
+    }
   }
 
   private checkBrowserSupportAndFeatureConfiguration(): void {